+++ lprint-zpl.h
@@ -11,6 +11,7 @@
 { "zpl_2inch-203dpi-tt",        "Zebra ZPL 2-inch/203dpi/Thermal-Transfer", "", NULL },
 { "zpl_2inch-300dpi-dt",        "Zebra ZPL 2-inch/300dpi/Direct-Thermal",   "", NULL },
 { "zpl_2inch-300dpi-tt",        "Zebra ZPL 2-inch/300dpi/Thermal-Transfer", "", NULL },
+{ "zpl_2inch-600dpi-tt",        "Zebra ZPL 2-inch/600dpi/Thermal-Transfer", "", NULL },
 { "zpl_4inch-203dpi-dt",        "Zebra ZPL 4-inch/203dpi/Direct-Thermal",
       "COMMAND SET:ZPL;", NULL },
 { "zpl_4inch-203dpi-dt-gx420d", "Zebra GX420d",
@@ -18,3 +19,4 @@
 { "zpl_4inch-203dpi-tt",        "Zebra ZPL 4-inch/203dpi/Thermal-Transfer", "", NULL },
 { "zpl_4inch-300dpi-dt",        "Zebra ZPL 4-inch/300dpi/Direct-Thermal",   "", NULL },
 { "zpl_4inch-300dpi-tt",        "Zebra ZPL 4-inch/300dpi/Thermal-Transfer", "", NULL },
+{ "zpl_4inch-600dpi-tt",        "Zebra ZPL 4-inch/600dpi/Thermal-Transfer", "", NULL },